+++ src/main/java/com/balsamic/sejongmalsami/repository/postgres/ExpRepository.java
@@ -1,6 +1,6 @@
 package com.balsamic.sejongmalsami.repository.postgres;
 
-import com.balsamic.sejongmalsami.object.Exp;
+import com.balsamic.sejongmalsami.object.postgres.Exp;
 import java.util.UUID;
 import org.springframework.data.jpa.repository.JpaRepository;
