+++ src/main/java/com/balsamic/sejongmalsami/repository/postgres/YeopjeonRepository.java
@@ -1,6 +1,6 @@
 package com.balsamic.sejongmalsami.repository.postgres;
 
-import com.balsamic.sejongmalsami.object.Yeopjeon;
+import com.balsamic.sejongmalsami.object.postgres.Yeopjeon;
 import java.util.UUID;
 import org.springframework.data.jpa.repository.JpaRepository;
